requirements_temp_utf8.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
psycopg2-binary==2.9.10
sqlalchemy==2.0.35
alembic==1.13.3
redis==5.1.1
asyncpg==0.29.0
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5
pycountry==24.6.1
torch==2.4.1
torch-geometric==2.6.1
scikit-learn==1.5.2
scipy==1.14.1
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1
google-cloud-bigquery==3.25.0
google-auth==2.35.0
google-cloud-core==2.4.1
httpx==0.27.2
requests==2.32.3
aiohttp==3.10.8
python-dotenv==1.0.1
pyyaml==6.0.2
python-multipart==0.0.12
python-dateutil==2.9.0.post0
pytz==2024.2
loguru==0.7.2
colorama==0.4.6
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
black==24.8.0
flake8==7.1.1
mypy==1.11.2
streamlit==1.39.0
altair==5.4.1
apscheduler==3.10.4
fastapi-cors==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
joblib==1.4.2
dill==0.3.8
beautifulsoup4==4.12.3
lxml==5.3.0
torch==2.4.1
torch-geometric==2.6.1
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
streamlit==1.39.0
plotly==5.24.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
supabase==2.8.1
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
requests==2.32.3